+++ packages/genesys-spark-components/src/components/stable/gux-dropdown-multi/gux-dropdown-multi.tsx
@@ -25,11 +25,11 @@ import translationResources from './i18n/en.json';
 import {
   getSearchOption,
   setInitialActiveOption,
-  getOptionDefaultSlot
+  getOptionDefaultSlot,
+  convertValueToArray
 } from '../gux-listbox/gux-listbox.service';
 import { GuxFilterTypes } from '../gux-dropdown/gux-dropdown.types';
 import { OnMutation } from '@utils/decorator/on-mutation';
-import { randomHTMLId } from '@utils/dom/random-html-id';
 /**
  * @slot - for a gux-listbox-multi containing gux-option-multi children
  */
@@ -176,7 +176,7 @@ export class GuxDropdownMulti {
    */
   @Method()
   getSelectedValues(): Promise<string[]> {
-    return Promise.resolve(this.value?.split(',') || []);
+    return Promise.resolve(convertValueToArray(this.value));
   }
 
   @Listen('keydown')
@@ -354,7 +354,7 @@ export class GuxDropdownMulti {
     const listboxOptionElements: HTMLGuxOptionElement[] = Array.from(
       this.root.querySelectorAll('gux-option-multi')
     );
-    const values = value ? value.split(',') : [];
+    const values = convertValueToArray(value);
 
     return listboxOptionElements.filter(element =>
       values.includes(element.value)
@@ -521,7 +521,7 @@ export class GuxDropdownMulti {
   }
 
   private renderTag(): JSX.Element {
-    const selectedValues = this.value?.split(',') || [];
+    const selectedValues = convertValueToArray(this.value);
     if (selectedValues.length) {
       return (
         <gux-dropdown-multi-tag
@@ -569,11 +569,9 @@ export class GuxDropdownMulti {
     }
   }
 
-  private uniquePopupID = randomHTMLId('gux-popup');
-
   private renderPopup(): JSX.Element {
     return (
-      <div slot="popup" id={this.uniquePopupID} class="gux-listbox-container">
+      <div slot="popup" class="gux-listbox-container">
         <slot />
       </div>
     ) as JSX.Element;
@@ -594,16 +592,14 @@ export class GuxDropdownMulti {
         slot="target"
       >
         {this.renderFilterInputField()}
-        <div
-          class="gux-field gux-field-combobox"
-          role="combobox"
-          aria-disabled={this.disabled}
+        <button
+          type="button"
+          class="gux-field gux-field-button"
+          disabled={this.disabled}
           onClick={this.fieldButtonClick.bind(this)}
           ref={el => (this.fieldButtonElement = el)}
           aria-haspopup="listbox"
           aria-expanded={this.expanded.toString()}
-          aria-controls={this.uniquePopupID}
-          tabIndex={this.disabled ? -1 : 0}
         >
           {this.renderTargetContent()}
           {this.renderTag()}
@@ -616,7 +612,7 @@ export class GuxDropdownMulti {
             screenreader-text={this.i18n('dropdown')}
             iconName="custom/chevron-down-small-regular"
           ></gux-icon>
-        </div>
+        </button>
       </div>
     ) as JSX.Element;
   }