+++ packages/genesys-spark-components/src/components/stable/gux-dropdown/gux-dropdown.tsx
@@ -438,13 +438,11 @@ export class GuxDropdown {
   }
 
   private renderOption(option: HTMLGuxOptionElement): JSX.Element {
-    let optionText = option.textContent;
+    let optionText = option.textContent.trim();
     if (hasSlot(option, 'subtext')) {
+      // TODO: use getOptionDefaultSlot(option)?.textContent.trim() once Stencil fix for assignedNodes test issue is in (v4.27.2)
       const subtext = option.querySelector('[slot=subtext]');
-      optionText = optionText.substring(
-        0,
-        optionText.length - subtext.textContent.length
-      );
+      optionText = optionText.replace(subtext.textContent, '');
     }
     return (
       <gux-truncate ref={el => (this.truncateElement = el)} dir="auto">
@@ -455,8 +453,15 @@ export class GuxDropdown {
 
   private renderIconOption(iconOption: HTMLGuxOptionIconElement): JSX.Element {
     let iconStyle = null;
+    let optionText = iconOption.textContent;
+
     if (iconOption.iconColor !== null) {
       iconStyle = { color: iconOption.iconColor };
+    }
+
+    if (hasSlot(iconOption, 'subtext')) {
+      const subtext = iconOption.querySelector('[slot=subtext]');
+      optionText = optionText.replace(subtext.textContent, '');
     }
     return (
       <span
@@ -472,7 +477,7 @@ export class GuxDropdown {
           size="small"
         ></gux-icon>
         <gux-truncate ref={el => (this.truncateElement = el)}>
-          {iconOption.textContent}
+          {optionText}
         </gux-truncate>
       </span>
     ) as JSX.Element;