+++ packages/genesys-spark-components/src/utils/dom/get-slot.ts
@@ -1,6 +1,12 @@
-export function getSlot(
-  element: HTMLElement,
-  slotName: string
-): HTMLSlotElement {
-  return element?.querySelector(`[slot=${slotName}]`);
+export function getSlot(element: HTMLElement, slotName: string): Element {
+  // This is a neater solution but our spec test do not support :scope selector
+  // return element?.querySelector(`:scope > *[slot=${slotName}]`);
+
+  for (const child of element.children) {
+    if (child.matches(`[slot=${slotName}]`)) {
+      return child;
+    }
+  }
+
+  return null;
 }