+++ packages/genesys-spark/scripts/snapshot.mjs
@@ -1,0 +1,42 @@
+#! /usr/bin/env node
+
+import dircompare from 'dir-compare';
+
+import path from 'path';
+import { fileURLToPath } from 'url';
+
+const filepath = fileURLToPath(import.meta.url);
+const dirname = path.dirname(filepath);
+const rootPath = path.resolve(dirname, '../');
+const options = { compareContent: true, noDiffSet: false };
+const dist = path.resolve(rootPath, 'dist');
+const snapshot = path.resolve(rootPath, 'snapshot');
+
+try {
+  const result = dircompare.compareSync(dist, snapshot, options);
+  const {
+    same,
+    differencesDirs,
+    differencesFiles,
+    totalDirs,
+    totalFiles,
+    diffSet
+  } = result;
+  if (!same) {
+    console.warn(`❌  'dist' and 'snapshot' folders are different.\n`);
+    console.error(`${differencesDirs}/${totalDirs} directories are distinct`);
+    console.error(`${differencesFiles}/${totalFiles} files are distinct`);
+    for (const diff of diffSet) {
+      if (diff.state !== 'equal') {
+        console.error(diff);
+      }
+    }
+    process.exit(1);
+  }
+  console.info(`✅  'dist' and 'snapshot' folders are identical.\n`);
+} catch {
+  console.warn(
+    `${rootPath}: ❌  'dist' folder missing. Run 'npm run build'.\n`
+  );
+  process.exit(1);
+}